libm: alm_main: warning message cleanup
Chages to cleanup warning messages.

AMD-Internal: [SWLCSG-545]

+++ include/libm_inlines_amd.h
@@ -361,8 +361,8 @@ static inline void splitexpf(float x, float logbase,
     else
       n = (int)(r - 0.5F);
 
-    r1 = x - n * logbaseof2_by_32_lead;
-    r2 =   - n * logbaseof2_by_32_trail;
+    r1 = x - (float)n * logbaseof2_by_32_lead;
+    r2 =   - (float)n * logbaseof2_by_32_trail;
 
     /* Set j = n mod 32:   5 mod 32 = 5,   -5 mod 32 = 27,  etc. */
     /* j = n % 32;
